fix #9: QR-codes consisting of numerics only are failed to scan
QR-code for numeric-only (e.g. "12" string http://qrcoder.ru/code/?12&10&0)
is not recognized with the test app.

Unit-test for this image causes AV.
The reason is in TDecoderResult.Destroy calls ByteSegments.Clear without
testing ByteSegments for nil. But ByteSegments is nil when QR-code is
numeric only.

Fix is to test ByteSegments for nil before calling ByteSegments.Clear.

Known issues: numerics of one digit is not still recognized.

Signed-off-by: Alexey Shumkin <Alex.Crezoff@gmail.com>